# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Water Compact Review API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Document {
  document_id: string;
  country: string;
  document_title: string;
  document_version: string;
  local_filename: string;
  official_source_url: string;
  sha256: string;
  file_size_bytes: number;
  page_count: number;
  page_numbering_basis: string;
  /** @nullable */
  pdf_creator?: string | null;
  /** @nullable */
  pdf_producer?: string | null;
  /** @nullable */
  pdf_creation_date?: string | null;
  /** @nullable */
  pdf_modified_date?: string | null;
  anchor_coordinate_system: string;
  generated_at_utc: string;
}

export interface Candidate {
  candidate_id: string;
  country: string;
  source_reference: string;
  compact_form: string;
  original_compact_text: string;
  standardized_candidate_reform_action: string;
  detailed_description: string;
  candidate_stream: string;
  inclusion_decision: string;
  reform_action_tag: string;
  count_in_reform_total: string;
  confidence_level: string;
  decision_rationale: string;
  water_security_pillar: string;
  solution_area: string;
  reform_type_tier_1: string;
  reform_type_tier_2: string;
  responsible_entity: string;
  /** @nullable */
  related_parent_id?: string | null;
  human_review_status: string;
  /** @nullable */
  reviewer_comment?: string | null;
  /** @nullable */
  final_validated_action?: string | null;
  reform_aspiration_status: string;
  criterion_assessment: string;
  review_count?: number;
  has_conflict?: boolean;
}

export interface AnchorSummary {
  anchor_id: string;
  candidate_id: string;
  document_id: string;
  anchor_role: string;
  anchor_type: string;
  /** @nullable */
  linked_candidate_id?: string | null;
  source_reference: string;
  page_start: number;
  page_end: number;
  matched_page: number;
  page_label: string;
  /** @nullable */
  section_or_table?: string | null;
  verbatim_source_text: string;
  normalized_source_text?: string;
  quote_sha256?: string;
  match_method?: string;
  /** @nullable */
  match_score?: number | null;
  /** @nullable */
  token_f1?: number | null;
  /** @nullable */
  token_precision?: number | null;
  /** @nullable */
  token_recall?: number | null;
  /** @nullable */
  sequence_similarity?: number | null;
  anchor_confidence: string;
  automated_validation_status?: string;
  human_review_status: string;
  /** @nullable */
  reviewer_comment?: string | null;
  created_at_utc: string;
}

export interface Rectangle {
  rectangle_id: string;
  anchor_id: string;
  page_number: number;
  x0: number;
  y0: number;
  x1: number;
  y1: number;
  x0_normalized: number;
  y0_normalized: number;
  x1_normalized: number;
  y1_normalized: number;
  page_width_points: number;
  page_height_points: number;
  coordinate_origin: string;
  coordinate_unit: string;
}

export type AnchorDetail = AnchorSummary & {
  rectangles: Rectangle[];
};

export type CandidateDetail = Candidate & {
  anchors: AnchorDetail[];
};

export interface CandidateListResponse {
  candidates: Candidate[];
  total: number;
  page: number;
  limit: number;
}

export interface Relationship {
  relationship_id: string;
  source_candidate_id: string;
  target_candidate_id: string;
  relationship_type: string;
  relationship_status: string;
  /** @nullable */
  source_anchor_id?: string | null;
  /** @nullable */
  target_anchor_id?: string | null;
  human_review_status: string;
  /** @nullable */
  reviewer_comment?: string | null;
}

export interface ReviewerCredentials {
  display_name: string;
  pin: string;
}

export interface ReviewerResponse {
  reviewer_id: string;
  display_name: string;
  /** Signed session token. Pass as Authorization Bearer on review submissions. */
  token: string;
}

export interface CandidateReview {
  review_id: number;
  candidate_id: string;
  reviewer_id: string;
  reviewer_name: string;
  inclusion_decision: string;
  /** @nullable */
  comment?: string | null;
  created_at: string;
}

export interface CandidateReviewsResponse {
  reviews: CandidateReview[];
  has_conflict: boolean;
}

export interface SubmitReviewRequest {
  inclusion_decision: string;
  /** @nullable */
  comment?: string | null;
}

export interface CountByKey {
  key: string;
  count: number;
}

export interface StatsResponse {
  total_candidates: number;
  total_documents: number;
  total_anchors: number;
  total_relationships: number;
  by_country: CountByKey[];
  by_stream: CountByKey[];
  by_inclusion_decision: CountByKey[];
  by_confidence_level: CountByKey[];
  by_anchor_confidence: CountByKey[];
  by_human_review_status: CountByKey[];
}

export type ListCandidatesParams = {
country?: string;
stream?: string;
inclusion_decision?: string;
confidence_level?: string;
anchor_confidence?: string;
reform_aspiration_status?: string;
reform_type_tier_1?: string;
human_review_status?: string;
/**
 * Free-text search across candidate_id, standardized action, actor, domain, source text
 */
search?: string;
page?: number;
limit?: number;
};

export type ListRelationshipsParams = {
candidateId?: string;
};